Optimization now requires a MultipleAlignment DS
